ROAD-10 frontend part

+++ src/app/components/dashboard/empresas/empresas.component.ts
@@ -2,6 +2,7 @@ import { Component, inject, OnDestroy, OnInit } from '@angular/core';
 import { PageEvent } from '@angular/material/paginator';
 import { EmpresaDTO } from '@models/DTOs/empresaDTO';
 import { MaterialModule } from '@modules/material.module';
+import { MessageDialogService } from '@services/message-dialog.service';
 import { ModalsService } from '@services/modals.service';
 import { DashBoardStore } from '@store/bashboard/dash-board-store';
 
@@ -13,8 +14,9 @@ import { DashBoardStore } from '@store/bashboard/dash-board-store';
 })
 export class EmpresasComponent implements OnInit {
 
-  public dashBoardStore = inject(DashBoardStore)
-  public displayedColumns: string[] = ['clave', 'nombreDes', 'linea', 'parentFleet', 'actions'];
+  public dashBoardStore = inject(DashBoardStore);
+  private messageDialogService = inject(MessageDialogService);
+  public displayedColumns: string[] = ['clave', 'descripcion', 'validador', 'dvr', 'actions'];
   private modalService = inject(ModalsService);
 
   ngOnInit(): void {
@@ -38,6 +40,9 @@ export class EmpresasComponent implements OnInit {
     this.modalService.openModal('empresa');
   }
   public async delete(id: number) {
+    const confirmation = await this.messageDialogService.confirmationMessage('¿Estás seguro que desea eliminar este registro?');
+    if (!confirmation) return;
+    
     await this.dashBoardStore.deleteEmpresa(id);
     this.modalService.closeModal();
   }
